fix: allow shortcode builder modal to scroll
It is possible for the shortcode builder to be taller than the viewport
for certain shortcodes like `[give_donor_wall]`. In these cases, the
user is unable to insert the shortcode because the button is outside the
viewport in a modal window with no way to access it.

This issue has been resolved by adding CSS that allows the modal window
to scroll vertically if it is too tall for the viewport. Shorter modal
windows will be centered vertically if they do not exceed 100% height.