fix: harden CSS of front-end donor wall
Some themes such as Twenty Fifteen have overreaching styles that are
affecting the appearance of the donor wall in negative ways especially
in terms of spacing and typography.

To improve the consistency of the donor wall appearance across themes, a
selective use of `!important` has been added to prevent themes from
breaking the donor wall layout and typography.