fix: ensure all comments are truncated near limit
Previously `str_pos()` was being used to truncate the excerpt, which is
problematic because the function looks forward for the next occurrence
in the haystack and may result in a `false` value if a space cannot be
found at the end of a comment.

This issue was resolved by using `strr_pos()` to look backwards from an
offset position to ensure that a comment is always truncated at the
space closest to the max character limit.