feature: normalize registered options (Fields API)
Prior, this was setting the option to be whatever kind array was given.
This created an ambiguous situation for consumers, since if an
associative array was given, the consumer would need to infer if the key
was the label or if the value was. Further, associative arrays are
limited in the sense that keys need to be unique. This is not the case
for options in a field. There can be duplicate values and duplicate
labels (though I would argue that the latter would be a bad user
experience, especially for those reliant on assistive technologies).

This commit creates a class called `FieldOption` which is constructed
with a value then a label (the latter of which is nullable). Numeric
arrays and associative arrays are both valid for registering options.
However, when they are set, they are normalized into this new format.
Further, providing an array of entries (or tuples) is now valid for
registering options. In that case, the order must match the constructor
of `FieldOption` (i.e. `$value, $label = null`).

Consumers can still need to handle labels potentially not being set,
however, it is now predictable as far as what is the value and what is
the label if both are set. Why not use the value as a fallback for the
label? This definitely could be an option, but I thought that a consumer
having the freedom to handle that how they want was the better
direction.

All consumers using the prior array options have been updated to use the
new `FieldOption` object.