feature: prepare class names before template
Concatenating the classes in the templates was getting pretty hairy so I
opted to take care of that in the `render` method. I included a simple
function (`Give\Helpers\Html::classNames`) which is a basic
implementation of the `classnames` JS library. It makes conditionally
setting class names easier. I suspect that we may need to set the
`$classAttribute` variable per template type and I think that the helper
will be very useful then.