refactor: expand GroupNode contract
I noticed that the `getNodeByName` didn’t exist as a method implemented
by `GroupNode`, so technically this code could fail if another class was
introduced that implemented `GroupNode`. Quick fix is to change the
instance the condition is checking for to `FieldCollection`, but I
thought that it would be more beneficial to add this method and a number
of others implemented by the `FieldCollection` class to the `GroupNode`
contract.