chore: use the classic runtime for React
Laravel Mix doesn’t support file-relative Babel config files. Well, it
actually likes to use a file-relative config as the project wide config
which is a bit strange. It seems to be silently obstinate about fixing
the matter. So this means we need to set the runtime for anything that’s
using React. Since we just recently changed it to automatic as an effect
of using LaraveL Mix, we know that we don’t actually need the automatic
runtime quite yet. In order to do this, we need to stop using
`.react()`. For simplicity sake, I just moved the entire Babel config
into the Mix file. While I was at it I cleaned up the Babel config,
since there were a few plugins that we didn’t really need.

I also moved the browserslist to the `package.json` file since that way
other tools can use it (e.g. autoprefixer). I also made a few other
changes that made sense in the `package.json`