fix: move currency switcher message outside of fieldset
The fieldset is useless visually — it just has the hidden fields for
currency switcher. It’s presence as a non-hidden element (for the sake
of its message) inside the grid layout messes up the gaps. Instead we
can hide it and move the message outside of it so that both it and the
message work well with the grid layout.