Bring back filter support for income goal stats
https://github.com/impress-org/givewp/commit/16e564b22272d326ddc12fccbaa162e49cdbe43b removed support for the filtered income value (`$goal_progress_stats['raw_actual']`) and this just makes that filterable again. It does this in a generalized way that lets anyone filter the income OR goal values in the context of this goal shortcode.

An added bonus here in this PR is that it passes the `$args` so that it can support even more complex queries for customization.

# Use-case

Customizing the current goal so that it's based on a date range. For instance, fundraising for specific versions of a free plugin so that once the fundraiser is over, the next version can have it's own separate income/goal without creating totally separate forms.