Switch the check for strauss.phar for better compatibility
I was experiencing an `sh` error when the `strauss.phar` file check was done using the `[[` syntax, which resulted in `strauss.phar` being downloaded every time that I ran `composer install` or `composer update`. This changeset switches to using the `test` command instead.

I ran a test with this `test` check in place on a fresh clone of this repo and the `.phar` file was downloaded and run successfully the first time. The second run of `composer install` avoided downloading the `.phar` and, again, it executed successfully. :man_dancing: