New: Support Slack-bot-based zip packaging requests
This tweaks the `generate-zip.yml` workflow so that it passes along some extra values to the `givewp-github-actions` zip workflow that enable versioned builds and Slack communication.

To support the versioned builds, a tweak to package.json is necessary so that we can read/write version number data in plugin files.