Chore: Add docblocks for necessary, but unused parameter (#6924)
The underlying `QueryBuilder` class supports the `$output` parameter which is passed to `DB::get_row` (`$wpdb->get_row()`). However, not all implementations of the `QueryBuilder`, such as the `ModelQueryBuilder`, support the different output types. The `$output` parameter is maintained for compatibility with the extended `QueryBuilder` class but is otherwise unused (enforced value) in some instances.

Co-authored-by: David GABISON <david.gabison@outlook.com>