Payment table: nonce must be built from translation of “form”
As an administrator, I would like to update donations with bulk action. The result is: 
We're unable to recognize your session. Please refresh the screen to try again; otherwise contact your website administrator for assistance.
The nonce verification failed.
In class-wp-list-table.php, the nonce is created from 'bulk-' . $this->_args['plural']. In french this id "formulaires". And the verification uses 'bulk-form'.