Return to the checkout form when the payment intent is created
When the payment intent cannot be create due to a Stripe error (Ex:  Your card's security code is incorrect.), there is an Exception. 
Some functions as handlePaymentIntentStatus() cannot access to status/id, and a new exception is created. There is no point in going any further, and we can return to the checkout form as in class-give-stripe-payment-method.php.