Fix: state hidden for Hungary
If the default country is Hungary, or if the donor has a saved address in Hungary, the 'State' field (named 'County' for Hungary) is field. But it's required, and the form doesn't answer with this message in the console:


If the donor change the country, then comes back to Hungary, the field is visible.